+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Features
 
+- Detach/Attach rapidly AWS policies to user, group or role with: `attach policy service=ec2 access=readonly group=sysadmin`. More info with `awless attach policy -h`
 - `awless ssh`: support SSH agent thanks to @justone
 - New `--port` flag for `awless ssh`: specifying non-standard SSH port thanks to @justone
 - Sync works on best effort now. Meaning it does not bail out when an error happens (most often it can be an access right issues on some AWS services)